move addCustomChain methods to network utils
